Builder being extended

+++ GuildGame/Assets/Core/Constants.cs
@@ -9,6 +9,6 @@ namespace com.Halcyon.Core
         public static readonly string LogsFolderPath = Path.Combine(Application.persistentDataPath, "Logs").Replace("\\", "/");
         public static readonly string SavesFolderPath = Path.Combine(Application.persistentDataPath, "Saves").Replace("\\", "/");
         public static readonly string DiscordInviteLink = "https://discord.gg/FnKBMfRqRb";
-        public static readonly int DefaultGridSize = 10;
+        public static readonly float DefaultGridSize = 10f;
     }
 }
